# API/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the documentation for the golang api of appolo-api",
        "title": "Appolo API",
        "contact": {
            "name": "API Support",
            "email": "pierre.saintsorny@gmail.com"
        },
        "version": "1"
    },
    "basePath": "/1.0",
    "paths": {
...
        "/binance/futures/order/buy": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Buy from binance futures using market order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_first_strategy",
                        "name": "strategy",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "0.01",
                        "name": "quantity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FuturesOrder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/binance/futures/order/sell": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "sell from binance futures using market order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_first_strategy",
                        "name": "strategy",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "0.01",
                        "name": "quantity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FuturesOrder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/binance/futures/position": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get Futures position informations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BTCUSDT",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "my_account",
                        "name": "account_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FuturesPosition"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
...
        "/strategy": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get one or multiple strategy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_first_strategy",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Strategy"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new strategy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BTCUSDT",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "my_first_strategy_BTCUSDT",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "1h",
                        "name": "interval",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "10",
                        "name": "leverage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "undefined",
                        "name": "account_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a strategy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_first_strategy",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "models.FuturesOrder": {
            "type": "object",
            "properties": {
                "commission": {
                    "type": "string"
                },
                "orderId": {
                    "type": "number"
                },
                "origQty": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "realizedPNL": {
                    "type": "string"
                },
                "side": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "integer"
                }
            }
        },
        "models.FuturesPosition": {
            "type": "object",
            "properties": {
                "entryPrice": {
                    "type": "string"
                },
                "leverage": {
                    "type": "string"
                },
                "positionAmt": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "unRealizedProfit": {
                    "type": "string"
                },
                "updateTime": {
                    "$ref": "#/definitions/models.intToTime"
                }
            }
        },
...
        "models.Strategy": {
            "type": "object",
            "properties": {
                "accountName": {
                    "type": "string"
                },
                "currencyName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interval": {
                    "type": "string"
                },
                "leverage": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.intToTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        }
    }
}